Remove stack navigator; move everything into tab nav and hide footer when needed
