+++ package.json
@@ -38,7 +38,6 @@
     "@react-navigation/drawer": "^6.3.1",
     "@react-navigation/elements": "^1.3.1",
     "@react-navigation/native": "^6.0.8",
-    "@react-navigation/native-stack": "^6.5.2",
     "@react-spring/native": "^9.5.5",
     "@realm/react": "^0.4.0",
     "@sayem314/react-native-keep-awake": "^1.1.0",
